끼니 전체조회 추가 ID로 식단 수정, 삭제 추가 식단 Date/Id 로직 클래스 분리 DailyMeal, EachMeal 클래스에 calculate 메서드 추가

+++ main_project/src/main/java/NutrientsCoders/main_project/dailymeal/dto/DailyMealResponseDto.java
@@ -46,7 +46,7 @@ public class DailyMealResponseDto {
       private String foodCategory2;
       private Long servingSize;
       private Double quantity;
-      private Long ratioEachKcal;
+      private Double ratioEachKcal;
       private Double ratioEachCarbo;
       private Double ratioEachProtein;
       private Double ratioEachFat;
